He creado todas las clases relacionadas a mi caso de uso Comprar Bono de Bocadillos, el diagrama de clases de uso y he creado el ApplicationDbContext y el ApplicationUser. (#16)

+++ src/AppForSEII2526.API/Models/Compra.cs
@@ -54,7 +54,7 @@ namespace AppForSEII2526.API.Models
 
         //PRECIO TOTAL
         [Precision(10, 2)]
-        public decimal PrecioTotal { get; set; }
+        public double PrecioTotal { get; set; }
 
         //NÚMERO DE BOCADILLOS
         [Required]
